move messages functions into separate files
