+++ src/component/files.ts
@@ -1,0 +1,64 @@
+import { assert, omit, pick } from "convex-helpers";
+import { paginator } from "convex-helpers/server/pagination";
+import { mergedStream, stream } from "convex-helpers/server/stream";
+import { nullable, partial } from "convex-helpers/validators";
+import { ObjectType } from "convex/values";
+import {
+  DEFAULT_MESSAGE_RANGE,
+  DEFAULT_RECENT_MESSAGES,
+  extractText,
+  isTool,
+} from "../shared.js";
+import {
+  paginationResultValidator,
+  vMessageStatus,
+  vMessageWithMetadata,
+  vSearchOptions,
+  vStepWithMessages,
+} from "../validators.js";
+import { api, internal } from "./_generated/api.js";
+import { Doc, Id } from "./_generated/dataModel.js";
+import {
+  action,
+  internalMutation,
+  internalQuery,
+  mutation,
+  MutationCtx,
+  query,
+  QueryCtx,
+} from "./_generated/server.js";
+import { schema, v } from "./schema.js";
+import { insertVector, searchVectors } from "./vector/index.js";
+import {
+  VectorDimension,
+  VectorDimensions,
+  VectorTableId,
+  vVectorId,
+} from "./vector/tables.js";
+import { paginationOptsValidator } from "convex/server";
+
+export const getFilesToDelete = query({
+  args: {
+    cursor: v.optional(v.string()),
+    limit: v.optional(v.number()),
+  },
+  handler: async (ctx, args) => {
+    const files = await paginator(ctx.db, schema)
+      .query("files")
+      .withIndex("refcount", (q) => q.eq("refcount", 0))
+      .paginate({
+        numItems: args.limit ?? 100,
+        cursor: args.cursor ?? null,
+      });
+    return {
+      files: files.page,
+      continueCursor: files.continueCursor,
+      isDone: files.isDone,
+    };
+  },
+  returns: v.object({
+    files: v.array(v.doc("files")),
+    continueCursor: v.string(),
+    isDone: v.boolean(),
+  }),
+});
